gpio/requirements.txt
attrs==18.2.0
autobahn==19.1.1
Automat==0.7.0
constantly==15.1.0
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
PyHamcrest==1.9.0
RPi.GPIO==0.6.5
six==1.12.0
Twisted==18.9.0
txaio==18.8.1
zope.interface==4.6.0
